Added property route
The list of properties of an entity is clickable and redirects
to the detailed view of the property (not programmed yet).
The routes have changed (renamed lab-mashup to entities).

The layout has changed and is work in progress.

The data model has changed. Properties have an own model and
are connected via relationships to the entity model.