!!! Move complete project to ReactJS !!!
ATTENTION: The complete project is ported to ReactJS.

Motivation: React (with flux) is much more flexible and performant in the
use case of VILLASweb. Ember is a good framework, but it has many tradeoffs
to its easy handling and feature completeness. Many work arounds had to be done
to get ember to work the way needed. Because ReactJS gives the developer much
more flexibility and feature choice, VILLASweb can be build much better.
The aim is to only depends on needed packages and be as performance as possible.

This new version still works with the current backend!

For library usage see package.json.

This first version contains the same base website as the old version but
changed back to legacy color scheme. Simulators are loaded from and can be added
to the backend as a proof-of-concept.